+++ AVInput/AVInput.cpp
@@ -31,7 +31,7 @@
 
 #define API_VERSION_NUMBER_MAJOR 1
 #define API_VERSION_NUMBER_MINOR 3
-#define API_VERSION_NUMBER_PATCH 1
+#define API_VERSION_NUMBER_PATCH 2
 
 #define HDMI 0
 #define COMPOSITE 1
@@ -325,7 +325,8 @@ string AVInput::currentVideoMode(bool &success)
 uint32_t AVInput::startInput(const JsonObject& parameters, JsonObject& response)
 {
     LOGINFOMETHOD();
-
+    
+    string sPortId = parameters["portId"].String();
     string sType = parameters["typeOfInput"].String();
     int portId = 0;
     int iType = 0;
@@ -333,7 +334,7 @@ uint32_t AVInput::startInput(const JsonObject& parameters, JsonObject& response)
     if (parameters.HasLabel("portId") && parameters.HasLabel("typeOfInput"))
     {
         try {
-            portId = parameters["portId"].Number();
+            portId = stoi(sPortId);
             iType = getTypeOfInput (sType);
         }catch (...) {
             LOGWARN("Invalid Arguments");
@@ -507,11 +508,12 @@ uint32_t AVInput::writeEDIDWrapper(const JsonObject& parameters, JsonObject& res
 {
     LOGINFOMETHOD();
 
-    int portId;
+    string sPortId = parameters["portId"].String();
+    int portId = 0;
     std::string message;
 
     if (parameters.HasLabel("portId") && parameters.HasLabel("message")) {
-        portId = parameters["portId"].Number();
+        portId = stoi(sPortId);
         message = parameters["message"].String();
     }
     else {
@@ -527,9 +529,10 @@ uint32_t AVInput::readEDIDWrapper(const JsonObject& parameters, JsonObject& resp
 {
     LOGINFOMETHOD();
 
+    string sPortId = parameters["portId"].String();
     int portId = 0;
     try {
-        portId = parameters["portId"].Number();
+        portId = stoi(sPortId);
     }catch (...) {
             LOGWARN("Invalid Arguments");
             returnResponse(false);
@@ -966,13 +969,14 @@ uint32_t AVInput::getSupportedGameFeatures(const JsonObject& parameters, JsonObj
 uint32_t AVInput::getGameFeatureStatusWrapper(const JsonObject& parameters, JsonObject& response)
 {
     string sGameFeature = "";
+    string sPortId = parameters["portId"].String();
     int portId = 0;
 
     LOGINFOMETHOD();
     if (parameters.HasLabel("portId") && parameters.HasLabel("gameFeature"))
     {
         try {
-            portId = parameters["portId"].Number();
+            portId = stoi(sPortId);
             sGameFeature = parameters["gameFeature"].String();
         }catch (...) {
             LOGWARN("Invalid Arguments");
@@ -1018,11 +1022,12 @@ uint32_t AVInput::getRawSPDWrapper(const JsonObject& parameters, JsonObject& res
 {
     LOGINFOMETHOD();
 
+    string sPortId = parameters["portId"].String();
     int portId = 0;
     if (parameters.HasLabel("portId"))
     {
         try {
-            portId = parameters["portId"].Number();
+            portId = stoi(sPortId);
         }catch (...) {
             LOGWARN("Invalid Arguments");
             returnResponse(false);
@@ -1047,11 +1052,12 @@ uint32_t AVInput::getSPDWrapper(const JsonObject& parameters, JsonObject& respon
 {
     LOGINFOMETHOD();
 
+    string sPortId = parameters["portId"].String();
     int portId = 0;
     if (parameters.HasLabel("portId"))
     {
         try {
-        portId = parameters["portId"].Number();
+            portId = stoi(sPortId);
         }catch (...) {
             LOGWARN("Invalid Arguments");
             returnResponse(false);
@@ -1149,12 +1155,13 @@ std::string AVInput::getSPD(int iPort)
 uint32_t AVInput::setEdidVersionWrapper(const JsonObject& parameters, JsonObject& response)
 {
     LOGINFOMETHOD();
+    string sPortId = parameters["portId"].String();
     int portId = 0;
     string sVersion = "";
     if (parameters.HasLabel("portId") && parameters.HasLabel("edidVersion"))
     {
         try {
-            portId = parameters["portId"].Number();
+            portId = stoi(sPortId);
             sVersion = parameters["edidVersion"].String();
         }catch (...) {
             LOGWARN("Invalid Arguments");
@@ -1202,13 +1209,14 @@ int AVInput::setEdidVersion(int iPort, int iEdidVer)
 
 uint32_t AVInput::getEdidVersionWrapper(const JsonObject& parameters, JsonObject& response)
 {
+    string sPortId = parameters["portId"].String();
     int portId = 0;
 
     LOGINFOMETHOD();
     if (parameters.HasLabel("portId"))
     {
         try {
-            portId = parameters["portId"].Number();
+            portId = stoi(sPortId);
         }
         catch (...) {
             LOGWARN("Invalid Arguments");