+++ Network/CHANGELOG.md
@@ -15,6 +15,14 @@ All notable changes to this RDK Service will be documented in this file.
 * Changes in CHANGELOG should be updated when commits are added to the main or release branches. There should be one CHANGELOG entry per JIRA Ticket. This is not enforced on sprint branches since there could be multiple changes for the same JIRA ticket during development. 
 
 * For more details, refer to [versioning](https://github.com/rdkcentral/rdkservices#versioning) section under Main README.
+## [1.2.0] - 2023-08-16
+### Added
+- Added new params in isConnectedToInternet for Query Internet Connectivity specific to Ipversion
+
+## [1.1.1] - 2023-08-16
+### Fixed
+- Remove the Cache Logic for getInterface Network Plugin api
+
 ## [1.1.0] - 2023-08-16
 ### Added
 - Added new method to configure PNI settings